# openApi.json
{"openapi":"3.0.0","paths":{"/api/v1/users/create":{"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"User created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseUserDto"}}}}},"tags":["Users"]}},"/api/v1/users/me":{"get":{"operationId":"UserController_currentUser","parameters":[],"responses":{"200":{"description":"Current user information","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseUserDto"}}}}},"tags":["Users"]}},"/api/v1/auth/sign-up":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"200":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["Auth"]}},"/api/v1/auth/sign-in":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"User logged in successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["Auth"]}},"/api/v1/role":{"post":{"operationId":"RoleController_createRole","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":""}},"tags":["Roles"]},"get":{"operationId":"RoleController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Roles"]}},"/api/v1/role/{name}":{"get":{"operationId":"RoleController_getRole","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Roles"]},"patch":{"operationId":"RoleController_updateRole","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["Roles"]},"delete":{"operationId":"RoleController_removeRole","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Roles"]}},"/api/v1/endpoint/all":{"get":{"operationId":"EndpointController_getAllRoutes","parameters":[],"responses":{"200":{"description":""}},"tags":["Endpoints"]}},"/api/v1/endpoint":{"post":{"operationId":"EndpointController_createEndpoint","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEndpointDto"}}}},"responses":{"201":{"description":""}},"tags":["Endpoints"]}},"/api/v1/permissions/allow":{"post":{"operationId":"PermissionsController_allowPermission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AllowPermissionDto"}}}},"responses":{"201":{"description":""}},"tags":["Permissions"]}}},"info":{"title":"Ecommerce API","description":"This is the API documentation for the Ecommerce API.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"authorization":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}},"required":["firstname","lastname","email","password","role"]},"ResponseRoleDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}},"required":["name","description"]},"ResponseUserDto":{"type":"object","properties":{"id":{"type":"number","example":1},"firstname":{"type":"string","example":"John"},"lastname":{"type":"string","example":"Doe"},"email":{"type":"string","example":"john.doe@example.com"},"isActive":{"type":"boolean","example":true},"role":{"$ref":"#/components/schemas/ResponseRoleDto"}},"required":["id","firstname","lastname","email","isActive","role"]},"SignUpDto":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["firstname","lastname","email","password"]},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","description":"JWT access token"},"user":{"$ref":"#/components/schemas/ResponseUserDto"}},"required":["accessToken","user"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string","default":"admin@gmail.com"},"password":{"type":"string","default":"Pass123."}},"required":["email","password"]},"CreateRoleDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}},"required":["name","description"]},"UpdateRoleDto":{"type":"object","properties":{"description":{"type":"string"}},"required":["description"]},"CreateEndpointDto":{"type":"object","properties":{"url":{"type":"string"},"method":{"type":"string"}},"required":["url","method"]},"AllowPermissionDto":{"type":"object","properties":{"roleName":{"type":"string"},"endpointId":{"type":"number"},"isAllow":{"type":"boolean"}},"required":["roleName","endpointId","isAllow"]}}}}