# src/auth/dto/auth-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { ResponseUserDto } from 'src/user/dto';

export class AuthResponseDto {
  @ApiProperty({
    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...',
    description: 'JWT access token',
  })
  accessToken: string;

  @ApiProperty({ type: () => ResponseUserDto })
  user: ResponseUserDto;
}